Refactor code to avoid circular dependency
